requirements.txt
fastapi==0.115.5
Flask==2.3.2
Flask_Cors==5.0.0
locust==2.16.1
urllib3==1.25.11
openai==1.7.0
Pillow==11.0.0
PyJWT==2.10.0
pymongo==4.3.3
meilisearch==0.31.5
PyPDF2==3.0.1
python-dotenv==1.0.1
bcrypt==4.0.1
Requests==2.32.3
uvicorn==0.32.1
redis==4.6.0
celery==5.2.7
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0
motor==3.3.2
meilisearch-python-async==1.0.2
